pyproject.toml
[project]
name = "hindi2spotify"
version = "0.1.0"
description = "Scrapes weekly top Hindi songs from JioSaavn and Gaana, matches them on Spotify, and updates a playlist daily."
requires-python = ">=3.11"
dependencies = [
    "spotipy>=2.23.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "thefuzz>=0.20.0",
    "python-Levenshtein>=0.23.0",
    "python-dotenv>=1.0.0",
]
